% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_discount_wts.R
\name{calc_discount_wts}
\alias{calc_discount_wts}
\title{Calculate Discount Weights}
\usage{
calc_discount_wts(discount_rate, num_cycles, cycle_length, assert = TRUE)
}
\arguments{
\item{discount_rate}{Annual discount rate expressed as a fraction
(e.g. 0.03).}

\item{num_cycles}{Integer giving the number of cycles to simulate.}

\item{cycle_length}{Numeric. Length of each cycle in years. Defaults to 1.}

\item{assert}{Logical. Boolean for whether to run full assertion of
inputs. This feature is useful when running probabilistic sensitivity
analysis (PSA).}
}
\value{
A numeric vector of discount weights of length \code{num_cycles + 1}.
}
\description{
Computes discount weights for a sequence of cycles given a discount rate
expressed per annum and the length of each cycle. These weights can be
multiplied by costs or QALYs to obtain discounted values.
}
\examples{
calc_discount_wts(
  discount_rate = 0.03,
  num_cycles = 5, 
  cycle_length = 1
)

}
