% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discount_rate_costs}
\alias{discount_rate_costs}
\title{Annual discount rate for costs}
\format{
A numeric value
}
\usage{
discount_rate_costs
}
\description{
Annual discount rate for costs
}
\keyword{datasets}
