% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discount_rate_QALYs}
\alias{discount_rate_QALYs}
\title{Annual discount rate for health outcomes}
\format{
A numeric value
}
\usage{
discount_rate_QALYs
}
\description{
Annual discount rate for health outcomes
}
\keyword{datasets}
