% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l_trans_probs}
\alias{l_trans_probs}
\title{List of transition probabilities and rates}
\format{
A list containing probabilities and rates for transitions between health states.
}
\usage{
l_trans_probs
}
\description{
List of transition probabilities and rates
}
\keyword{datasets}
