% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_indi_features}
\alias{m_indi_features}
\title{Matrix of individual features (age and sex)}
\format{
A matrix with two columns: "age" and "sex"
}
\usage{
m_indi_features
}
\description{
Matrix of individual features (age and sex)
}
\keyword{datasets}
