% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{num_cycles}
\alias{num_cycles}
\title{Time horizon of the simulation in cycles}
\format{
An integer
}
\usage{
num_cycles
}
\description{
Time horizon of the simulation in cycles
}
\keyword{datasets}
