% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_probsV.R
\name{update_probsV}
\alias{update_probsV}
\title{Update Transition Probabilities}
\usage{
update_probsV(
  v_states_names,
  v_occupied_state,
  l_trans_probs,
  v_time_in_state,
  assert = TRUE
)
}
\arguments{
\item{v_states_names}{Character vector of all state names in the order used
throughout the model.}

\item{v_occupied_state}{Character vector of current state for each
individual.}

\item{l_trans_probs}{Named list of transition probabilities and rate ratios;
see \code{update_probsV} for details.}

\item{v_time_in_state}{Numeric vector indicating the time spent in the
current state.}

\item{assert}{Logical. Boolean for whether to run full assertion of
inputs. This feature is useful when running probabilistic sensitivity
analysis (PSA).}
}
\value{
A matrix of transition probabilities with rows corresponding to
individuals and columns corresponding to the health states listed in
\code{v_states_names}. Each row sums to one.
}
\description{
Computes a transition probability matrix for each individual in the cohort
given the current health state, time spent in that state and a list of
baseline probabilities and modifiers. Probabilities are adjusted
for mortality using rate ratios and time in state. The resulting matrix
has one row per individual and one column for each possible next state.
}
\examples{
v_states_names <- c("H", "S1", "S2", "D")
v_occupied_state <- c("H", "S1")
l_trans_probs <- list(
  p_HD = 0.005,
  p_HS1 = 0.15,
  p_S1H = 0.5,
  p_S1S2 = 0.105,
  p_S1D = 1 - exp(-3 * (-log(1-0.005))),
  p_S2D = 1 - exp(-10 * (-log(1-0.005))), rp_S1 = 0.2,
  rp_S2 = 0.29
)
v_time_in_state <- c(1, 2)

m_probs <- update_probsV(
  v_states_names,
  v_occupied_state,
  l_trans_probs,
  v_time_in_state
 )

rowSums(m_probs)

}
