% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_cost_coeffs}
\alias{v_cost_coeffs}
\title{Vector of cost regression coefficients}
\format{
A numeric vector
}
\usage{
v_cost_coeffs
}
\description{
Vector of cost regression coefficients
}
\keyword{datasets}
