% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_starting_states}
\alias{v_starting_states}
\title{The starting health state for all individuals}
\format{
A character vector
}
\usage{
v_starting_states
}
\description{
The starting health state for all individuals
}
\keyword{datasets}
