% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_states_costs}
\alias{v_states_costs}
\title{Vector of costs for each health state (no treatment)}
\format{
A named numeric vector
}
\usage{
v_states_costs
}
\description{
Vector of costs for each health state (no treatment)
}
\keyword{datasets}
