% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_states_costs1}
\alias{v_states_costs1}
\title{Vector of costs for each health state (treatment 1)}
\format{
A named numeric vector
}
\usage{
v_states_costs1
}
\description{
Vector of costs for each health state (treatment 1)
}
\keyword{datasets}
