% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_states_names}
\alias{v_states_names}
\title{Vector of the model health state names}
\format{
A character vector
}
\usage{
v_states_names
}
\description{
Vector of the model health state names
}
\keyword{datasets}
