% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_states_utilities}
\alias{v_states_utilities}
\title{Vector of utilities for each health state (no treatment)}
\format{
A named numeric vector
}
\usage{
v_states_utilities
}
\description{
Vector of utilities for each health state (no treatment)
}
\keyword{datasets}
