% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_states_utilities2}
\alias{v_states_utilities2}
\title{Vector of utilities for each health state (treatment 2)}
\format{
A named numeric vector
}
\usage{
v_states_utilities2
}
\description{
Vector of utilities for each health state (treatment 2)
}
\keyword{datasets}
