% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_util_coeffs}
\alias{v_util_coeffs}
\title{Vector of utility regression coefficients}
\format{
A numeric vector
}
\usage{
v_util_coeffs
}
\description{
Vector of utility regression coefficients
}
\keyword{datasets}
