% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_util_t_decs}
\alias{v_util_t_decs}
\title{Vector of state-specific utility decrements over time}
\format{
A numeric vector
}
\usage{
v_util_t_decs
}
\description{
Vector of state-specific utility decrements over time
}
\keyword{datasets}
